pyproject.toml
[tool.poetry]
name = "my-study-api"
version = "0.1.0"
description = "An api to create quizzes and flashcards"
authors = ["MatheusBLopes <matheusblopes10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
supermemo2 = "^2.0.0"
pydantic = "^2.7.1"
sqlmodel = "^0.0.18"
pydantic-settings = "^2.2.1"
uvicorn = "^0.29.0"
alembic = "^1.13.1"
fastapi-sqlalchemy = "^0.2.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
